seperated worker class
